Extract delete related logic to tree_delete mod
